Core/Quest: Corrected type of emote field in QuestOfferReward to be signed
Port From (https://github.com/TrinityCore/TrinityCore/commit/27931b45c681f7f84928bad557882350731df3dc)

+++ Source/Game/Entities/Creature/Gossip.cs
@@ -529,7 +529,7 @@ namespace Game.Misc
             offer.SuggestedPartyMembers = quest.SuggestedPlayers;
 
             for (uint i = 0; i < SharedConst.QuestEmoteCount && quest.OfferRewardEmote[i] != 0; ++i)
-                offer.Emotes.Add(new QuestDescEmote(quest.OfferRewardEmote[i], quest.OfferRewardEmoteDelay[i]));
+                offer.Emotes.Add(new QuestDescEmote((uint)quest.OfferRewardEmote[i], quest.OfferRewardEmoteDelay[i]));
 
             offer.QuestFlags[0] = (uint)quest.Flags;
             offer.QuestFlags[1] = (uint)quest.FlagsEx;
